# docs/swagger.yaml
definitions:
  api_errors.APIErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  model.Account:
    properties:
      id:
        type: string
    type: object
  model.Android:
    properties:
      android_group_id:
        type: string
      id:
        type: string
    type: object
  model.AndroidGroup:
    properties:
      account_id:
        type: string
      id:
        type: string
    type: object
  model.CreateAccountResponse:
    properties:
      account:
        $ref: '#/definitions/model.Account'
      android:
        $ref: '#/definitions/model.Android'
      android_group:
        $ref: '#/definitions/model.AndroidGroup'
    type: object
  model.ErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/api_errors.APIErrorResponse'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  title: account-service documentation
  version: 0.0.1
paths:
  /accounts:
    post:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreateAccountResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Create a new account and associated android group and android. All
        newly created entities are returned.
      tags:
      - accounts
  /accounts/{id}:
    delete:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Delete an account
      tags:
      - accounts
    get:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get details of an account
      tags:
      - accounts
  /android_groups/{id}/androids:
    post:
      parameters:
      - description: Android Group ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Android'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Create a new account
swagger: "2.0"